# Optimize normalize_flat_keys